Fix link to discussion page in FAQ
There is an awkward problem for API based content that can contain
links to internal router pages. You cannot set a nuxt-link component
inside a string and have nuxt interpret it as an internal link.
(See https://medium.com/@wearethreebears/handle-api-driven-content-links-in-nuxt-js-fe2e31ecbeeb)

This problem is solved here by adding the nuxt-interpolation library
that intercepts all click events and pushes the location to the nuxt
router if it is an internal link and adds a noopener attribute to
external links.